+++ extracting_tests.xml
@@ -2,9 +2,6 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Chrome tests"  parallel="methods" thread-count="3">
 
-    <listeners>
-        <listener class-name="listeners.TestListener"></listener> (Add your listener class path here)
-    </listeners>
     <parameter name="browser" value="CHROME" />
     <test name="Chrome"  >
 
@@ -17,7 +14,7 @@
             </run>
         </groups>
         <classes>
-            <class name="tests.Spice"/>
+            <class name="tests.Parallel"/>
 
         </classes>
